Rename and remove Next dependency
