+++ package.json
@@ -8,9 +8,7 @@
     "@farcaster/fishery": "^2.2.3",
     "@farcaster/hub-nodejs": "^0.10.21",
     "ethers": "^6.10.0",
-    "next": "13.5.6",
     "react": "^18",
-    "react-dom": "^18.2.0",
     "viem": "^2.5.0"
   },
   "scripts": {
@@ -31,7 +29,6 @@
     "@changesets/changelog-github": "^0.4.8",
     "@changesets/cli": "^2.26.2",
     "@types/jest": "^29.5.11",
-    "@types/react": "^18.2.48",
     "jest": "^29.7.0",
     "prettier": "^3.1.1",
     "prettier-plugin-tailwindcss": "^0.5.9",
