+++ src/core/farcasterTypes.ts
@@ -36,3 +36,16 @@ export function convertToFrame(json: any) {
     },
   };
 }
+
+export type Button = {
+  label: string;
+  action?: 'post' | 'post_redirect';
+};
+
+export type FrameMetadata = {
+  image: string;
+  buttons?: [Button, ...Button[]];
+  post_url?: string;
+  refresh_period?: number;
+};
+
