+++ src/index.ts
@@ -2,7 +2,7 @@
 const version = '0.3.0';
 
 export { version };
-export { createFrameResponse } from './core/createFrameResponse';
+export { getFrameHtml } from './core/getFrameHtml';
 export { getFrameAccountAddress } from './core/getFrameAccountAddress';
 export { getFrameMetadata } from './core/getFrameMetadata';
 export { getFrameMessage } from './core/getFrameMessage';
